+++ src/Services/Board/Board.Application/Dtos/SearchServiceModels/Events/UserActionEvents/BoardItemActionEvent.cs
@@ -4,9 +4,9 @@ namespace Elastic.Application.IntegrationEvents.Events.UserActionEvents
 {
 	public record BoardItemActionEvent : UserActionEvent
 	{
-		public int BoardItemId { get; init; }
-		public int BoardId { get; init; }
-		public BoardItemActionEvent(string userId, int boardItemId, int boardId, string description = "") : base(userId, description)
+		public long BoardItemId { get; init; }
+		public long BoardId { get; init; }
+		public BoardItemActionEvent(string userId, long boardItemId, long boardId, string description = "") : base(userId, description)
 		{
 			this.BoardItemId = boardItemId;
 			this.BoardId = boardId;
