+++ src/Services/Board/Board.Application/Dtos/SearchServiceModels/Events/UserActionEvents/CommentActionEvent.cs
@@ -4,9 +4,9 @@ namespace Elastic.Application.IntegrationEvents.Events.UserActionEvents
 {
 	public record CommentActionEvent : UserActionEvent
 	{
-		public int CommentId { get; init; }
-		public int BoardItemId { get; init; }
-		public CommentActionEvent(string username, int commentId, string description = "") : base(username, description)
+		public long CommentId { get; init; }
+		public long BoardItemId { get; init; }
+		public CommentActionEvent(string username, long commentId, string description = "") : base(username, description)
 		{
 			this.CommentId = commentId;
 		}
