+++ src/Services/Board/Board.Application/Dtos/SearchServiceModels/Events/UserActionEvents/DeleteBoardActionEvent.cs
@@ -4,7 +4,7 @@ namespace Elastic.Application.IntegrationEvents.Events.UserActionEvents
 {
 	public record DeleteBoardActionEvent : BoardActionEvent
 	{
-		public DeleteBoardActionEvent(string username, int boardId, string description = "") : base(username, boardId, description)
+		public DeleteBoardActionEvent(string username, long boardId, string description = "") : base(username, boardId, description)
 		{
 		}
 	}
