+++ src/Services/Board/Board.Application/Dtos/SearchServiceModels/Events/UserActionEvents/DeleteBoardItemActionEvent.cs
@@ -4,7 +4,7 @@ namespace Elastic.Application.IntegrationEvents.Events.UserActionEvents
 {
 	public record DeleteBoardItemActionEvent : BoardItemActionEvent
 	{
-		public DeleteBoardItemActionEvent(string username, int boardItemId, int boardId, string description = "") : base(username, boardItemId, boardId, description)
+		public DeleteBoardItemActionEvent(string username, long boardItemId, int boardId, string description = "") : base(username, boardItemId, boardId, description)
 		{
 		}
 	}
