+++ src/Services/Board/Board.Application/Dtos/SearchServiceModels/Events/UserActionEvents/DeleteCommentActionEvent.cs
@@ -4,7 +4,7 @@ namespace Elastic.Application.IntegrationEvents.Events.UserActionEvents
 {
 	public record DeleteCommentActionEvent : CommentActionEvent
 	{
-		public DeleteCommentActionEvent(string username, int commentId, string description = "") : base(username, commentId, description)
+		public DeleteCommentActionEvent(string username, long commentId, string description = "") : base(username, commentId, description)
 		{
 		}
 	}
