+++ src/Services/Elastic/Elastic.Application/Dao/EntityDao.cs
@@ -9,7 +9,7 @@ namespace Elastic.Application.Dao
 	public record EntityDao : BaseDao
 	{
 		//Like Board Id, BorderItem Id, and 
-		public int EntityId { get; set; }
+		public long EntityId { get; set; }
 		public string Creator { get; set; }
 		public DateTime CreationDate { get; set; }
 		public string Keyword { get; set; }
