+++ src/Services/Elastic/Elastic.Application/IntegrationEvents/Events/UserActionEvents/CreateThumbUpActionEvent.cs
@@ -9,7 +9,7 @@ namespace Elastic.Application.IntegrationEvents.Events.UserActionEvents
 {
 	public record CreateThumbUpActionEvent : CommentActionEvent
 	{
-		private CreateThumbUpActionEvent(string userName, int commentId, string description = "") : base(userName, commentId, description)
+		private CreateThumbUpActionEvent(string userName, long commentId, string description = "") : base(userName, commentId, description)
 		{
 		}
