+++ src/Services/Elastic/Elastic.Application/IntegrationEvents/Events/UserActionEvents/DeleteCommentActionEvent.cs
@@ -8,7 +8,7 @@ namespace Elastic.Application.IntegrationEvents.Events.UserActionEvents
 {
 	public record DeleteCommentActionEvent : CommentActionEvent
 	{
-		public DeleteCommentActionEvent(string username, int commentId, string description = "") : base(username, commentId, description)
+		public DeleteCommentActionEvent(string username, long commentId, string description = "") : base(username, commentId, description)
 		{
 		}
 	}
