+++ src/Services/Elastic/Elastic.Application/IntegrationEvents/Events/UserActionEvents/UpdateBoardActionEvent.cs
@@ -10,7 +10,7 @@ namespace Elastic.Application.IntegrationEvents.Events.UserActionEvents
 	public record UpdateBoardActionEvent : BoardActionEvent
 	{
 		public BoardDao Board { get; set; }
-		private UpdateBoardActionEvent(string userName, int boardId, string description = "") : base(userName, boardId, description)
+		private UpdateBoardActionEvent(string userName, long boardId, string description = "") : base(userName, boardId, description)
 		{
 
 		}
