+++ src/Services/Elastic/Elastic.Application/IntegrationEvents/Events/UserActionEvents/UpdateCommentActionEvent.cs
@@ -10,7 +10,7 @@ namespace Elastic.Application.IntegrationEvents.Events.UserActionEvents
 	public record UpdateCommentActionEvent : CommentActionEvent
 	{
 		public CommentDao Comment { get; set; }
-		private UpdateCommentActionEvent(string userName, int commentId, string description = "") : base(userName, commentId, description)
+		private UpdateCommentActionEvent(string userName, long commentId, string description = "") : base(userName, commentId, description)
 		{
 		}
 		public UpdateCommentActionEvent(CommentDao comment, string description = "") : this(comment.Creator, comment.EntityId, description)
